# docs/swagger.yaml
basePath: /api/v1/messages
definitions:
  model.MessageDto:
    properties:
      content:
        type: string
      phone:
        type: string
    type: object
  model.MessageView:
    properties:
      content:
        type: string
      id:
        type: string
      messageId:
        type: string
      phone:
        type: string
      sent:
        type: boolean
      sentAt:
        type: string
    type: object
info:
  contact: {}
  description: Gin API with Swagger documentation Message service for Insider
  title: Insider Message Service
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      description: Save a new message with phone number and content. It will be sent
        automatically later.
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.MessageDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create message
      tags:
      - Message
  /sent:
    get:
      consumes:
      - application/json
      description: Retrieves all messages from the database that have been marked
        as sent
      produces:
      - application/json
      responses:
        "200":
          description: List of delivered messages
          schema:
            items:
              $ref: '#/definitions/model.MessageView'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List delivered (sent) messages
      tags:
      - Message
  /start:
    post:
      consumes:
      - application/json
      description: Starts a background process that sends unsent messages every 2
        minutes
      produces:
      - application/json
      responses:
        "200":
          description: status message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start automatic message sending
      tags:
      - Message
  /stop:
    post:
      consumes:
      - application/json
      description: Stops the background process that sends unsent messages every 2
        minutes
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Stop automatic message sending
      tags:
      - Message
swagger: "2.0"